# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Secret Broker API
  version: 4.2.0
  description: mTLS Secret Broker for AI clients. AI never sees plaintext keys.
servers:
  - url: https://broker.example.com
    description: Aliyun production
  - url: https://broker-bk.example.com
    description: Tencent production (failover)
  - url: https://localhost:8443
    description: Local development with a trusted test CA
security:
  - mtls: []
  - bearerAuth: []
  - sessionCookie: []
...
paths:
...
  /api/v2/operations:
    post:
      tags:
        - operations-v2
      summary: Create a policy-bound typed operation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationCreate"
      responses:
        "202":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v2/auth/webauthn/begin:
    post:
      tags:
        - auth-v2
      summary: Begin phishing-resistant WebAuthn authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - client
              properties:
                client:
                  type: string
                  maxLength: 128
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebAuthnOptions"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v2/auth/webauthn/finish:
    post:
      tags:
        - auth-v2
      summary: Verify WebAuthn assertion and create a short browser session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebAuthnFinish"
      responses:
        "200":
          description: Authenticated; session is returned only as a secure cookie
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v2/me/webauthn/registration/begin:
    post:
      tags:
        - auth-v2
      summary: Begin security-key registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                label:
                  type: string
                  maxLength: 80
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebAuthnOptions"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v2/me/webauthn/registration/finish:
    post:
      tags:
        - auth-v2
      summary: Verify and store a security-key credential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebAuthnFinish"
      responses:
        "201":
          description: Credential registered
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v2/approvals:
    post:
      tags:
        - approvals-v2
      summary: Create a fully bound approval request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalCreate"
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Approval is not required
    get:
      tags:
        - approvals-v2
      summary: List approval requests visible to the requester or approver
      responses:
        "200":
          description: Approval list
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v2/browser/otp/claim:
    post:
      tags:
        - browser-v2
      summary: Claim one approved OTP for a bound browser document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrowserOtpClaim"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrowserOtpClaimResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v2/browser/otp/finish:
    post:
      tags:
        - browser-v2
      summary: Finalize a single-use browser OTP claim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrowserOtpFinish"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Claim expired or already consumed
  /api/v2/devices/enroll/begin:
    post:
      tags:
        - devices-v2
      summary: Begin proof-of-possession device enrollment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceEnrollmentBegin"
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceEnrollmentChallenge"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v2/devices/enroll/finish:
    post:
      tags:
        - devices-v2
      summary: Finish device enrollment with a supported hardware-backed signature
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceEnrollmentFinish"
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v2/devices:
    get:
      tags:
        - devices-v2
      summary: List visible devices without public keys or secrets
      responses:
        "200":
          description: Device list
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    mtls:
      type: mutualTLS
      description: mTLS client certificate
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key (mb_live_xxx)
    sessionCookie:
      type: apiKey
      in: cookie
      name: broker_session
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        detail:
          type: string
        request_id:
          type: string
        retry_after_ms:
          type: integer
        docs:
          type: string
...
    OperationCreate:
      type: object
      additionalProperties: false
      required:
        - provider
        - operation_id
        - account_ref
        - environment
        - typed_parameters
      properties:
        provider:
          type: string
          pattern: ^[a-z][a-z0-9_-]{1,63}$
        operation_id:
          type: string
          pattern: ^[a-z][a-z0-9_.-]{1,127}$
        account_ref:
          type: string
          minLength: 1
          maxLength: 128
        environment:
          type: string
          enum:
            - development
            - staging
            - production
        typed_parameters:
          type: object
        approval_request_id:
          type: string
          format: uuid
          description: Single-use approval bound to this exact request
        otp:
          $ref: "#/components/schemas/OtpRequirement"
    ApprovalCreate:
      type: object
      additionalProperties: false
      required:
        - provider
        - operation_id
        - account_ref
        - environment
        - typed_parameters
      properties:
        provider:
          type: string
          pattern: ^[a-z][a-z0-9_-]{1,63}$
        operation_id:
          type: string
          pattern: ^[a-z][a-z0-9_.-]{1,127}$
        account_ref:
          type: string
          minLength: 1
          maxLength: 128
        environment:
          type: string
          enum:
            - development
            - staging
            - production
        typed_parameters:
          type: object
...
    Approval:
      type: object
      additionalProperties: false
      required:
        - id
        - requester
        - provider
        - operation_id
        - account_ref
        - environment
        - resource_ref
        - required_approvals
        - approvals
        - status
        - created_at
        - expires_at
      properties:
        id:
          type: string
          format: uuid
        requester:
          type: string
        provider:
          type: string
        operation_id:
          type: string
        account_ref:
          type: string
        environment:
          type: string
        resource_ref:
          type: string
        required_approvals:
          type: integer
          minimum: 1
          maximum: 10
        approvals:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - approved_by
              - approved_at
            properties:
              approved_by:
                type: string
              approved_at:
                type: string
                format: date-time
        status:
          type: string
          enum:
            - pending
            - approved
            - rejected
            - expired
            - consumed
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    OtpRequirement:
      type: object
      additionalProperties: false
      required:
        - device_id
        - sim_binding
        - challenge
        - recipient_hint
      properties:
        device_id:
          type: string
          format: uuid
        sim_binding:
          type: string
          maxLength: 128
        challenge:
          type: string
          maxLength: 256
        recipient_hint:
          type: string
          maxLength: 64
        template_group:
          type: string
          maxLength: 64
        ttl_seconds:
          type: integer
          minimum: 15
          maximum: 120
    Operation:
      type: object
      required:
        - id
        - provider
        - operation_id
        - status
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
        operation_id:
          type: string
        account_ref:
          type: string
        environment:
          type: string
        status:
          type: string
          enum:
            - waiting
            - received
            - consuming
            - completed
            - failed
            - expired
            - revoked
        result:
          type: object
        error:
          type: object
        created_at:
          type: string
          format: date-time
    DeviceEnrollmentBegin:
      type: object
      additionalProperties: false
      required:
        - label
        - platform
      properties:
        label:
          type: string
          maxLength: 80
        platform:
          type: string
          enum:
            - android
            - windows
            - linux
            - ios
            - browser-worker
        capabilities:
          type: array
          items:
            type: string
    DeviceEnrollmentChallenge:
      type: object
      required:
        - enrollment_id
        - challenge
        - expires_at
      properties:
        enrollment_id:
          type: string
          format: uuid
        challenge:
          type: string
        expires_at:
          type: string
          format: date-time
    DeviceEnrollmentFinish:
      type: object
      additionalProperties: false
      required:
        - enrollment_id
        - signature_algorithm
        - public_key_pem
        - signature
      properties:
        enrollment_id:
          type: string
          format: uuid
        signature_algorithm:
          type: string
          enum:
            - ed25519
            - p256-sha256
        public_key_pem:
          type: string
          description: PEM encoded Ed25519 or NIST P-256 public key
        signature:
          type: string
          description: Base64url signature over the enrollment challenge
    Device:
      type: object
      required:
        - id
        - label
        - platform
        - signature_algorithm
        - state
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        platform:
          type: string
        signature_algorithm:
          type: string
          enum:
            - ed25519
            - p256-sha256
        state:
          type: string
          enum:
            - active
            - suspended
            - revoked
        capabilities:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
...
    BrowserOtpClaim:
      type: object
      additionalProperties: false
      required:
        - provider
        - account_ref
        - origin
        - tab_id
        - frame_id
        - document_id
      properties:
        provider:
          type: string
          enum:
            - aliyun
            - tencent
        account_ref:
          type: string
          pattern: ^[a-z0-9][a-z0-9._:-]{0,127}$
        origin:
          type: string
          enum:
            - https://account.aliyun.com
            - https://cloud.tencent.com
        tab_id:
          type: integer
          minimum: 0
        frame_id:
          type: integer
          const: 0
        document_id:
          type: string
          minLength: 1
          maxLength: 256
    BrowserOtpClaimResult:
      type: object
      additionalProperties: false
      required:
        - type
        - provider
        - account_ref
        - origin
        - tab_id
        - frame_id
        - document_id
        - expires_at_ms
        - code
        - receipt
      properties:
        type:
          type: string
          const: approved-otp
        provider:
          type: string
        origin:
          type: string
        account_ref:
          type: string
        tab_id:
          type: integer
        frame_id:
          type: integer
        document_id:
          type: string
        expires_at_ms:
          type: integer
        code:
          type: string
          pattern: ^[0-9]{4,10}$
        receipt:
          type: string
    BrowserOtpFinish:
      type: object
      additionalProperties: false
      required:
        - receipt
        - completed
      properties:
        receipt:
          type: string
          minLength: 1
          maxLength: 128
        completed:
          type: boolean
    WebAuthnOptions:
      type: object
      additionalProperties: false
      required:
        - flow_id
        - options
      properties:
        flow_id:
          type: string
          format: uuid
        options:
          type: object
          additionalProperties: true
    WebAuthnFinish:
      type: object
      additionalProperties: false
      required:
        - flow_id
        - response
      properties:
        flow_id:
          type: string
          format: uuid
        response:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...